added origin concept to Api State
Apis now have a location and an origin.

The origin can be set via `.origin(val)` and is available in all route
and close State contexts via `this.origin`.

Origin should not contain any routing data but rather carries meta
information about the request (which protocol requested it etc.). This
information should be used to transform the output to be viable for the
origin of the request (e.g. HTTP vs WebSocket).